% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualFunctions.R
\name{countryGoalsMap}
\alias{countryGoalsMap}
\title{Produces Leaflet Map that shows how many goals each country has scored}
\usage{
countryGoalsMap()
}
\value{
a leaflet map of countries and their respective goals
}
\description{
Produces Leaflet Map that shows how many goals each country has scored
}
