% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualFunctions.R
\name{getCountryGoals}
\alias{getCountryGoals}
\title{Helper Function which returns the finalized dataframe of countries and goals they have scored in the worldcup}
\usage{
getCountryGoals()
}
\value{
a dataframe
}
\description{
Helper Function which returns the finalized dataframe of countries and goals they have scored in the worldcup
}
