% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualFunctions.R
\name{getCountryLatLong}
\alias{getCountryLatLong}
\title{Helper Function which returns the finalized dataframe of countries lat/long values}
\usage{
getCountryLatLong()
}
\value{
a dataframe
}
\description{
Helper Function which returns the finalized dataframe of countries lat/long values
}
