% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{getMatchInfo}
\alias{getMatchInfo}
\title{Returns the information for a given match in the world cup}
\usage{
getMatchInfo(id)
}
\arguments{
\item{id}{id of the world cup match}
}
\value{
string with information about the specific match
}
\description{
Returns the information for a given match in the world cup
}
