% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{getPlayerInfo}
\alias{getPlayerInfo}
\title{Returns the information for a given player name}
\usage{
getPlayerInfo(name)
}
\arguments{
\item{name}{A player's name at the world cup}
}
\value{
dataframe with numeric/char information about player
}
\description{
Returns the information for a given player name
}
