% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualFunctions.R
\name{getStadiumsInfo}
\alias{getStadiumsInfo}
\title{Produces a list of attendance/coordinates at each stadium that was used in the Finals}
\usage{
getStadiumsInfo()
}
\value{
A list of attendance/coordinates at each stadium that was used in the Finals
}
\description{
Produces a list of attendance/coordinates at each stadium that was used in the Finals
}
