% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Temp.R
\name{getsummarised}
\alias{getsummarised}
\title{Gets the summarized world cup data frame from the list}
\usage{
getsummarised()
}
\value{
A data frame
}
\description{
Gets the summarized world cup data frame from the list
}
