% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Temp.R
\name{tidyWorldCup}
\alias{tidyWorldCup}
\title{Produces a tidy dataframe of world cup data for the user}
\usage{
tidyWorldCup()
}
\value{
A data Frame
}
\description{
Produces a tidy dataframe of world cup data for the user
}
