% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualFunctions.R
\name{topGoalScorers}
\alias{topGoalScorers}
\title{Produces Bar Graph with Top Goal Scorers in WC Competitions}
\usage{
topGoalScorers(num)
}
\value{
a bar plot
}
\description{
Produces Bar Graph with Top Goal Scorers in WC Competitions
}
