% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{winningTeam}
\alias{winningTeam}
\title{Returns which team won the world cup given year input}
\usage{
winningTeam(year)
}
\arguments{
\item{year}{The year of the world cup}
}
\value{
A string
}
\description{
Returns which team won the world cup given year input
}
