% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{yearsWon}
\alias{yearsWon}
\title{Returns the years won for a given country input}
\usage{
yearsWon(country)
}
\arguments{
\item{country}{A country at the world cup}
}
\value{
dataframe of numeric values
}
\description{
Returns the years won for a given country input
}
