% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getSummarised}
\alias{getSummarised}
\title{Gets the summarized world cup data frame from the list}
\usage{
getSummarised()
}
\value{
A data frame
}
\description{
Gets the summarized world cup data frame from the list
}
