% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Temp.R
\name{tidyWorldCup}
\alias{tidyWorldCup}
\title{Produces a list of tidy data frames of world cup data for the user}
\usage{
tidyWorldCup()
}
\value{
A list of data frames
}
\description{
Produces a list of tidy data frames of world cup data for the user
}
